clipbutler_proxy/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
google-cloud-storage>=2.16.0
google-cloud-aiplatform>=1.50.0
vertexai>=1.50.0
sqlalchemy>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
